move ClassmateContext to SPI package, since it occurs in BootstrapContext
it was previously in an internal package, and leaked onto the SPI
